Fix/invalid version (#17)
* java.net のバージョン変更漏れを修正

* update the node modules

* Skip delete proccess at no deletion target

* Fix the Zulu Collector failed

+++ backend/OpenJdkReleaseInformationCollector/lambda/src/main/java/com/github/poad/openjdk/finder/backend/service/JavaVersionService.java
@@ -34,6 +34,9 @@ public class JavaVersionService {
     }
 
     public void delete(List<String> ids) {
+        if (ids.isEmpty()) {
+            return;
+        }
         jdbi.withExtension(JavaVersionRepository.class, repository -> {
             repository.delete(distribution, ids);
             return null;
